use strict mode clean architecture
