+++ .github/workflows/codeql.yaml
@@ -26,7 +26,7 @@ jobs:
         language: [ 'go' ]
 
     steps:
-    - name: Checkout repository
+    - name: Checkout infrastructure
       uses: actions/checkout@v4
 
     - name: Initialize CodeQL
