+++ app/user/controllers/rpc/handler.go
@@ -20,22 +20,23 @@ import (
 	"context"
 
 	"github.com/west2-online/DomTok/app/user/controllers/rpc/pack"
-	"github.com/west2-online/DomTok/app/user/domain"
+	"github.com/west2-online/DomTok/app/user/domain/model"
+	"github.com/west2-online/DomTok/app/user/usecase"
 	"github.com/west2-online/DomTok/kitex_gen/user"
 )
 
 // UserHandler 实现 idl 中定义的 RPC 接口
 type UserHandler struct {
-	useCase domain.UserUseCase
+	useCase usecase.UserUseCase
 }
 
-func NewUserHandler(useCase domain.UserUseCase) *UserHandler {
+func NewUserHandler(useCase usecase.UserUseCase) *UserHandler {
 	return &UserHandler{useCase: useCase}
 }
 
 func (h *UserHandler) Register(ctx context.Context, req *user.RegisterRequest) (r *user.RegisterResponse, err error) {
 	r = new(user.RegisterResponse)
-	u := &domain.User{
+	u := &model.User{
 		UserName: req.Username,
 		Password: req.Password,
 		Email:    req.Email,
@@ -54,7 +55,7 @@ func (h *UserHandler) Register(ctx context.Context, req *user.RegisterRequest) (
 func (h *UserHandler) Login(ctx context.Context, req *user.LoginRequest) (r *user.LoginResponse, err error) {
 	r = new(user.LoginResponse)
 
-	u := &domain.User{
+	u := &model.User{
 		UserName: req.Username,
 		Password: req.Password,
 	}
