+++ app/user/controllers/rpc/pack/pack.go
@@ -17,12 +17,12 @@ limitations under the License.
 package pack
 
 import (
-	"github.com/west2-online/DomTok/app/user/domain"
+	model2 "github.com/west2-online/DomTok/app/user/domain/model"
 	"github.com/west2-online/DomTok/kitex_gen/model"
 )
 
 // BuildUser 将 entities 定义的 User 实体转换成 idl 定义的 RPC 交流实体，类似 dto
-func BuildUser(u *domain.User) *model.UserInfo {
+func BuildUser(u *model2.User) *model.UserInfo {
 	return &model.UserInfo{
 		UserId: u.Uid,
 		Name:   u.UserName,
