+++ app/user/domain/repository/interface.go
@@ -14,14 +14,18 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package domain
+package repository
 
-// User 用于在 handler --- use case --- repository 之间传递数据的实体类
-// 目的是方便 use case 操作对应的业务
-type User struct {
-	Uid      int64
-	UserName string
-	Password string
-	Email    string
-	Phone    string
+import (
+	"context"
+
+	"github.com/west2-online/DomTok/app/user/domain/model"
+)
+
+// domain中的 repository 表示 service / use case 所依赖的外部资源，比如数据库、缓存等
+
+// UserDB 表示持久化存储接口 (或者也可以叫做 DBPort)
+type UserDB interface {
+	IsUserExist(ctx context.Context, username string) (bool, error)
+	CreateUser(ctx context.Context, user *model.User) error
 }
