+++ app/user/usecase/usecase.go
@@ -17,18 +17,27 @@ limitations under the License.
 package usecase
 
 import (
-	"github.com/west2-online/DomTok/app/user/domain"
-	"github.com/west2-online/DomTok/app/user/service"
+	"context"
+
+	"github.com/west2-online/DomTok/app/user/domain/model"
+	"github.com/west2-online/DomTok/app/user/domain/repository"
+	"github.com/west2-online/DomTok/app/user/domain/service"
 )
+
+// UserUseCase 接口应该不应该定义在 domain 中，这属于 use case 层
+type UserUseCase interface {
+	RegisterUser(ctx context.Context, user *model.User) (uid int64, err error)
+	Login(ctx context.Context, user *model.User) (*model.User, error)
+}
 
 // useCase 实现了 domain.UserUseCase
 // 只会以接口的形式被调用, 所以首字母小写改为私有类型
 type useCase struct {
-	db  domain.UserDB
+	db  repository.UserDB
 	svc *service.UserService
 }
 
-func NewUserCase(db domain.UserDB, svc *service.UserService) domain.UserUseCase {
+func NewUserCase(db repository.UserDB, svc *service.UserService) *useCase {
 	return &useCase{
 		db:  db,
 		svc: svc,
